fix: Correct cursor alignment on high-DPI displays
This commit fixes a bug where the mouse cursor position was not correctly aligned with the objects on the canvas, particularly on high-DPI (Retina) screens.

The bug was caused by two issues introduced during the high-DPI refactoring:
1. The `getMousePos` function did not scale the mouse's CSS pixel coordinates to the canvas's physical pixel coordinates.
2. The `resizeCanvas` function contained an erroneous `ctx.scale()` call which caused inconsistent rendering.

This patch addresses both issues:
- The `getMousePos` function now correctly multiplies the mouse coordinates by the device pixel ratio.
- The incorrect `ctx.scale()` call has been removed from `resizeCanvas`.
